Split user_input module into grids and model module
