+++ src/lcm/__init__.py
@@ -1,4 +1,5 @@
 from lcm import mark
-from lcm.user_input import DiscreteGrid, LinspaceGrid, LogspaceGrid, Model
+from lcm.grids import DiscreteGrid, LinspaceGrid, LogspaceGrid
+from lcm.model import Model
 
 __all__ = ["mark", "Model", "LinspaceGrid", "LogspaceGrid", "DiscreteGrid"]
